relocate the logic that retrieves distribution and lattice defaults to defaults_helper.py
they were originally stored in generalFunctions.py. It makes sense to move this logic to the defaults helper file. No functionality changes. Just location
